% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributionAverage.r
\name{averageContribution}
\alias{averageContribution}
\title{Retrieve average contribution for one or more fit indeces}
\usage{
averageContribution(da.object, fit.functions = NULL)
}
\arguments{
\item{da.object}{dominanceAnalysis object}

\item{fit.functions}{name of the fit indeces to retrieve. If NULL, all fit indeces will be retrieved}
}
\value{
a list. Key corresponds to fit-index and the value is vector,
                with average contribution for each variable
}
\description{
Retrieve the average contribution for a dominanceAnalysis object.
The average contribution defines general dominance.
}
