% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centerOnGroup.r
\name{centerOnGroup}
\alias{centerOnGroup}
\title{Center the variables on groups mean
Returns a dataframe with variables groups means as x.mean
and centered variables a x.centered}
\usage{
centerOnGroup(x, g)
}
\arguments{
\item{x}{dataframe}

\item{g}{grouping factor}
}
\value{
New dataframe
}
\description{
Center the variables on groups mean
Returns a dataframe with variables groups means as x.mean
and centered variables a x.centered
}
