% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daUtils.r
\name{checkDominanceAnalysis}
\alias{checkDominanceAnalysis}
\title{Check if the given object have the dominanceAnalysis class}
\usage{
checkDominanceAnalysis(x)
}
\arguments{
\item{x}{an object}
}
\value{
boolean TRUE if x is a dominaceAnalysis object, raises an error otherwise
}
\description{
Stop execution if object isn't a dominanceAnalysis object
}
\keyword{internal}
