% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributionByLevel.r
\name{contributionByLevel}
\alias{contributionByLevel}
\title{Retrieve contribution by level for one or more fit indeces}
\usage{
contributionByLevel(da.object, fit.functions = NULL)
}
\arguments{
\item{da.object}{dominanceAnalysis object}

\item{fit.functions}{name of the fit indeces to retrieve. If NULL, all fit indeces will be retrieved}
}
\value{
a list. Key corresponds to fit-index and the value is a matrix, with contribution of each variable
                by level
}
\description{
Retrieve the contribution by level for a given dominanceAnalysis object.
This contribution defines conditional dominance.
}
