% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daFitFunctions.r
\name{da.dynlm.fit}
\alias{da.dynlm.fit}
\title{Provides coefficient of determination for \code{dynlm} models.}
\usage{
da.dynlm.fit(original.model, newdata = NULL, ...)
}
\arguments{
\item{original.model}{Original fitted model}

\item{newdata}{Data used in update statement}

\item{...}{ignored}
}
\value{
A function described by \link{using-fit-indices} description for interface
}
\description{
Uses \eqn{R^2} (coefficient of determination) as fit index
}
\seealso{
Other fit indices: 
\code{\link{da.betareg.fit}()},
\code{\link{da.clm.fit}()},
\code{\link{da.glm.fit}()},
\code{\link{da.lm.fit}()},
\code{\link{da.lmWithCov.fit}()},
\code{\link{da.lmerMod.fit}()},
\code{\link{da.mlmWithCov.fit}()}
}
\concept{fit indices}
