\name{da.glm.fit}
\alias{da.glm.fit}
\title{Uses fit indexes provided by Azen and Traxel(2009)
- r2.m: McFadden(1974)
- r2.n: Nagelkerke(1991)
- r2.e: Estrella(1998)}
\usage{
  da.glm.fit(...)
}
\description{
  Uses fit indexes provided by Azen and Traxel(2009) -
  r2.m: McFadden(1974) - r2.n: Nagelkerke(1991) - r2.e:
  Estrella(1998)
}

