% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daFitFunctions.r
\name{da.lmerMod.fit}
\alias{da.lmerMod.fit}
\title{Provides fit indices for hierarchical linear models, based on Luo and Azen (2013).}
\usage{
da.lmerMod.fit(data, null.model, ...)
}
\arguments{
\item{data}{complete data set containing the variables in the model}

\item{null.model}{needed for HLM models}

\item{...}{ignored}
}
\description{
Provides fit indices for hierarchical linear models, based on Luo and Azen (2013).
}
\references{
\itemize{
\item Luo, W., & Azen, R. (2012). Determining Predictor Importance in Hierarchical Linear Models Using Dominance Analysis. Journal of Educational and Behavioral Statistics, 38(1), 3-31. doi:10.3102/1076998612458319
}
}
\seealso{
Other fit indices: \code{\link{da.dynlm.fit}},
  \code{\link{da.glm.fit}}, \code{\link{da.lm.fit}},
  \code{\link{da.lmWithCov.fit}},
  \code{\link{da.mlmWithCov.fit}}
}
\concept{fit indices}
