% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daFitFunctions.r
\name{da.lmWithCov.fit}
\alias{da.lmWithCov.fit}
\title{Provides coefficient of determination for linear models, using covariance/correlation matrix.}
\usage{
da.lmWithCov.fit(base.cov, ...)
}
\arguments{
\item{base.cov}{variance/covariance matrix}

\item{...}{ignored}
}
\description{
Uses \eqn{R^2} (coefficient of determination)
See \code{\link{lmWithCov}}
}
\seealso{
Other fit indices: \code{\link{da.dynlm.fit}},
  \code{\link{da.glm.fit}}, \code{\link{da.lm.fit}},
  \code{\link{da.lmerMod.fit}},
  \code{\link{da.mlmWithCov.fit}}
}
\concept{fit indices}
