% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daAverageContributionByLevel.r
\name{daAverageContributionByLevel}
\alias{daAverageContributionByLevel}
\title{Calculates the average contribution by level for a daRawResult object}
\usage{
daAverageContributionByLevel(x)
}
\arguments{
\item{x}{daRawResults object}
}
\value{
list, with key named as a fit index
              and values are matrix, with the average
              contribution of each variable on every level
}
\description{
Calculates the average contribution by level for a daRawResult object
}
\keyword{internal}
