% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daCompleteDominance.r
\name{daCompleteDominance}
\alias{daCompleteDominance}
\title{Matrix of Complete dominance of one variable over another
Uses daRawResults as input}
\usage{
daCompleteDominance(daRR)
}
\description{
Matrix of Complete dominance of one variable over another
Uses daRawResults as input
}
\seealso{
Other dominance matrices: \code{\link{daConditionalDominance}},
  \code{\link{daGeneralDominance}}
}
\keyword{internal}
