% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daGeneralDominance.r
\name{daGeneralDominance}
\alias{daGeneralDominance}
\title{Matrix of General dominance of one variable over another
Uses daRawResults as input}
\usage{
daGeneralDominance(daRR)
}
\description{
Matrix of General dominance of one variable over another
Uses daRawResults as input
}
\seealso{
Other dominance matrices: \code{\link{daCompleteDominance}},
  \code{\link{daConditionalDominance}}
}
\concept{dominance matrices}
\keyword{internal}
