% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daSubmodels.r
\name{daSubmodels}
\alias{daSubmodels}
\title{Returns all the submodels derived from full models.}
\usage{
daSubmodels(x, constants = NULL, terms = NULL)
}
\arguments{
\item{x}{regression class (lm or lmer)}

\item{constants}{vector of constants}

\item{terms}{vector of terms. By default, obtained using the formula}
}
\value{
list with elements level, pred.matrix, predictors, response, constants
}
\description{
You could set some variables as constantes, limiting the number of models.
Includes, by default, the null model
}
\keyword{internal}
