\name{dominanceAnalysis}
\alias{dominanceAnalysis}
\title{Dominance Analysis
Based on Azen and Bodescu(1993) and all their derivations}
\usage{
  dominanceAnalysis(x, constants = c(),
    fit.functions = "default", data = NULL, ...)
}
\arguments{
  \item{x}{lm, glm or another regression based ll}

  \item{data.frame,}{if you don't provide a lm object}
}
\description{
  Dominance Analysis Based on Azen and Bodescu(1993) and
  all their derivations
}

