% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominanceMatrix.r
\name{dominanceMatrix.data.frame}
\alias{dominanceMatrix.data.frame}
\title{Calculates the dominance for a given matrix.
Just transform the data.frame to matrix an call \link{dominanceMatrix.matrix}}
\usage{
\method{dominanceMatrix}{data.frame}(x, undefined.value = 0.5,
  type = NULL, fit.function = NULL, ...)
}
\description{
Calculates the dominance for a given matrix.
Just transform the data.frame to matrix an call \link{dominanceMatrix.matrix}
}
\keyword{internal}
