% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominanceMatrix.r
\name{dominanceMatrix.dominanceAnalysis}
\alias{dominanceMatrix.dominanceAnalysis}
\title{Retrieve a dominance matrix for a dominanceAnalysis object}
\usage{
\method{dominanceMatrix}{dominanceAnalysis}(x, undefined.value = 0.5,
  type = NULL, fit.function = NULL, ...)
}
\arguments{
\item{x}{a dominanceAnalysis object}

\item{undefined.value}{unused}

\item{type}{type of dominance matrix. Could be complete, conditional or general}

\item{fit.function}{specific fit function to retrieve. If not specified, retrieve the first
method according to fit.functions}
}
\value{
A matrix representing dominance. 1 represented domination of the row variable
        over the column variable, 0 dominance of the column over the row variable.
        Undefined dominance is represented by \code{undefined.value} parameter
}
\description{
This methods allows a common interface to retrieve all dominance matrices
from dominanceAnalysis objects
}
\keyword{internal}
