% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominanceMatrix.r
\name{dominanceMatrix.matrix}
\alias{dominanceMatrix.matrix}
\title{Calculates the dominance for a given matrix}
\usage{
\method{dominanceMatrix}{matrix}(x, undefined.value = 0.5, type = NULL,
  fit.function = NULL, ...)
}
\arguments{
\item{x}{matrix with fit indexes. Should have colnames}

\item{undefined.value}{value when no dominance can be established}

\item{type}{unused}

\item{fit.function}{unused}
}
\value{
A matrix representing dominance. 1 represented domination of the row variable
        over the column variable, 0 dominance of the column over the row variable.
        Undefined dominance is represented by \code{undefined.value} parameter
}
\description{
Dominance is defined between a pair of variables if all values on a given
variable are higher than for other variable.
}
\keyword{internal}
