\name{dominanceMatrix}
\alias{dominanceMatrix}
\title{Calculates the dominance for a given matrix
Dominance is defined between a pair of variables if all values on a given
variable are higher than for other variable}
\usage{
  dominanceMatrix(x)
}
\description{
  Calculates the dominance for a given matrix Dominance is
  defined between a pair of variables if all values on a
  given variable are higher than for other variable
}

