% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daUtils.r
\name{getEqualRowId}
\alias{getEqualRowId}
\title{Return the index of row equals to r on m}
\usage{
getEqualRowId(m, r)
}
\arguments{
\item{m}{matrix}

\item{r}{row}
}
\description{
Return the index of row equals to r on m
}
\keyword{internal}

