% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmR2.r
\name{lmmR2}
\alias{lmmR2}
\title{Calculates several measures of fit for Linear Mixel Models
based on Lou and Azen (2013) text.
Models could be lmer or lme models}
\usage{
lmmR2(m.null, m.full)
}
\arguments{
\item{m.null}{Null model (only with random intercept effects)}

\item{m.full}{Full model}
}
\value{
lmmR2 class
}
\description{
Calculates several measures of fit for Linear Mixel Models
based on Lou and Azen (2013) text.
Models could be lmer or lme models
}
