\name{lmWithCov}
\alias{lmWithCov}
\title{Uses covariance/correlation matrix for OLS
We use for testing, comparing the results of papers with
this package, but useful anyway}
\usage{
  lmWithCov(f, x)
}
\arguments{
  \item{f}{formula}

  \item{x}{correlation/covariance matrix}
}
\description{
  Uses covariance/correlation matrix for OLS We use for
  testing, comparing the results of papers with this
  package, but useful anyway
}

