% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominanceAnalysis.r
\name{print.dominanceAnalysis}
\alias{print.dominanceAnalysis}
\title{Print the results of a dominanceAnalysis object}
\usage{
\method{print}{dominanceAnalysis}(x, ...)
}
\value{
an object of type \code{dominanceAnalysis}
}
\description{
Print the results of a dominanceAnalysis object
}
\keyword{internal}
