% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmR2.r
\name{print.lmmR2}
\alias{print.lmmR2}
\title{Print method for lmmR2 object, that retrieves the summary.}
\usage{
\method{print}{lmmR2}(x, ...)
}
\arguments{
\item{x}{lmmR2 object}

\item{...}{extra arguments for print}
}
\value{
an lmmR2 object
}
\description{
Print method for lmmR2 object, that retrieves the summary.
}
\keyword{internal}
