% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bootDominanceAnalysis.r
\name{print.summary.bootDominanceAnalysis}
\alias{print.summary.bootDominanceAnalysis}
\title{Print a summary.bootDominanceAnalysis object}
\usage{
\method{print}{summary.bootDominanceAnalysis}(x, round.digits = 3, ...)
}
\arguments{
\item{x}{a \code{\link{summary.bootDominanceAnalysis}} object}

\item{round.digits}{Number of decimal places to round results}

\item{...}{further arguments passed to print method}
}
\value{
an object \code{\link{summary.bootDominanceAnalysis}}
}
\description{
Print a summary.bootDominanceAnalysis object
}
\keyword{internal}
