% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dominanceAnalysis.r
\name{print.summary.dominanceAnalysis}
\alias{print.summary.dominanceAnalysis}
\title{Print a summary.dominanceAnalysis object}
\usage{
\method{print}{summary.dominanceAnalysis}(x, round.digits = 3, ...)
}
\arguments{
\item{x}{a \code{\link{summary.dominanceAnalysis}} object}

\item{round.digits}{Number of decimal places to round results}

\item{...}{further arguments passed to print method}
}
\value{
a summary.dominanceAnalysis object
}
\description{
Print a summary.dominanceAnalysis object
}
\keyword{internal}
