% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmR2.r
\name{print.summary.lmmR2}
\alias{print.summary.lmmR2}
\title{Print method for  lmmR2 models summary}
\usage{
\method{print}{summary.lmmR2}(x, ...)
}
\arguments{
\item{x}{summary.lmmR2 object}

\item{...}{unused}
}
\description{
Print method for  lmmR2 models summary
}
