% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankUsingMatrix.r
\name{rankUsingMatrix}
\alias{rankUsingMatrix}
\title{Dominance of each factor over others.
Dominance requires that a variable have higher values on all submodels.
This method allows to visualize those relation.}
\usage{
rankUsingMatrix(x)
}
\arguments{
\item{x}{a square matrix}
}
\value{
a vector of string, each one showing what factors dominates others
}
\description{
Dominance of each factor over others.
Dominance requires that a variable have higher values on all submodels.
This method allows to visualize those relation.
}
\keyword{internal}
