% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daUtils.r
\name{replaceTermsInString}
\alias{replaceTermsInString}
\title{Replace terms by name using the terms definition}
\usage{
replaceTermsInString(string, replacement)
}
\arguments{
\item{string}{string to be updated}

\item{replacement}{string with replacement for strings. values are replaced by names}
}
\description{
Replace terms by name using the terms definition
}
