\name{retrieveAllModels}
\alias{retrieveAllModels}
\title{Returns all the formulas to a base model
using predictor groups
Factors are considered basic groups}
\usage{
  retrieveAllModels(x, groups = c())
}
\value{
  list list of names models, with names of predictors
}
\description{
  Returns all the formulas to a base model using predictor
  groups Factors are considered basic groups
}

