% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bootDominanceAnalysis.r
\name{summary.bootDominanceAnalysis}
\alias{summary.bootDominanceAnalysis}
\title{Summary for bootDominanceAnalysis.}
\usage{
\method{summary}{bootDominanceAnalysis}(object, fit.functions = NULL,
  ...)
}
\arguments{
\item{object}{a \code{\link{bootDominanceAnalysis}} object}

\item{fit.functions}{name of the fit indeces to retrieve. If NULL, all fit indeces will be retrieved}

\item{...}{ignored}
}
\description{
Summary for bootDominanceAnalysis.
}
\keyword{internal}
