% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dominanceAnalysis.r
\name{summary.dominanceAnalysis}
\alias{summary.dominanceAnalysis}
\title{Summary for a \code{\link{dominanceAnalysis}} object}
\usage{
\method{summary}{dominanceAnalysis}(object, ...)
}
\arguments{
\item{object}{a \code{\link{dominanceAnalysis}} object}

\item{...}{unused}
}
\description{
Summary for a \code{\link{dominanceAnalysis}} object
}
