\name{bootDominanceAnalysis}
\alias{bootDominanceAnalysis}
\title{Bootstrap analysis for Dominance Analysis
Based on Azen and Bodescu(2006)}
\usage{
  bootDominanceAnalysis(x, R, constants = c(),
    fit.functions = "default", null.model = NULL, ...)
}
\description{
  Bootstrap analysis for Dominance Analysis Based on Azen
  and Bodescu(2006)
}

