% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootDominanceAnalysis.r
\name{bootDominanceAnalysis}
\alias{bootDominanceAnalysis}
\title{Bootstrap analysis for Dominance Analysis}
\usage{
bootDominanceAnalysis(x, R, constants = c(), fit.functions = "default",
  null.model = NULL, ...)
}
\arguments{
\item{x}{lm, glm, lmer model}

\item{R}{number of bootstrap resamples}

\item{constants}{vector of variables to remain unchanged between models}

\item{fit.functions}{list of functions which provides fit indexes for model.}

\item{null.model}{for mixel models, null model against to test the submodels}

\item{...}{Other arguments provided to lm or lmer (not implemented yet)}
}
\description{
Bootstrap procedure as presented on Azen and Bodescu(2003).
Provides the expected level of dominance of variable \eqn{X_i} over \eqn{X_j},
as the degree to which the pattern found on sample is reproduced on the
bootstrap samples.
Use \code{\link{summary.bootDominanceAnalysis}} to get a nice formatted
data.frame
}
\examples{
\dontrun{
lm.1<-lm(Employed~.,longley)
da.boot<-bootDominanceAnalysis(lm.1,R=1000)
summary(da.boot)
}
}
\seealso{
\code{\link{summary.bootDominanceAnalysis}}
}
