% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daFitFunctions.r
\name{da.glm.fit}
\alias{da.glm.fit}
\title{Provides fit indexes for GLM models, based on Azen and Traxel(2009)}
\usage{
da.glm.fit(...)
}
\value{
A list with several fit indexes
\item{r2.m}{McFadden(1974)}
\item{r2.cs}{Cox and Snell(1989). Use as a reference, because don't have 1 as upper bound}
\item{r2.n}{Nagelkerke(1991), that corrects the upper bound of Cox and Snell(1989) index }
\item{r2.e}{Estrella(1998)}
}
\description{
Provides fit indexes for GLM models, based on Azen and Traxel(2009)
}
\references{
\itemize{
\item Azen, R., & Traxel, N. (2009). Using Dominance Analysis to Determine Predictor Importance in Logistic Regression. Journal of Educational and Behavioral Statistics, 34(3), 319-347. doi:10.3102/1076998609332754
\item Nagelkerke, N. J. D. (1991). A note on a general definition of the coefficient of determination. Biometrika, 78(3), 691–692. doi:10.1093/biomet/78.3.691
}
}
\seealso{
Other fit indexes: \code{\link{da.lm.fit}},
  \code{\link{da.lmWithCov.fit}},
  \code{\link{da.lmerMod.fit}},
  \code{\link{da.mlmWithCov.fit}}
}
