% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daFitFunctions.r
\name{da.glm.fit}
\alias{da.glm.fit}
\title{Provides fit indexes for GLM models, based on Azen and Traxel(2009)}
\usage{
da.glm.fit(...)
}
\value{
A list with several fit indexes 
\item{r2.m}{McFadden(1974)}
\item{r2.n}{Nagelkerke(1991)}
\item{r2.e}{Estrella(1998)}
}
\description{
Provides fit indexes for GLM models, based on Azen and Traxel(2009)
}
\references{
Azen, R., & Traxel, N. (2009). Using Dominance Analysis to Determine Predictor Importance in Logistic Regression. Journal of Educational and Behavioral Statistics, 34(3), 319-347. doi:10.3102/1076998609332754
}

