% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daFitFunctions.r
\name{da.mlmWithCov.fit}
\alias{da.mlmWithCov.fit}
\title{Provides coefficient of determination for multivariate models.}
\usage{
da.mlmWithCov.fit(base.cov, ...)
}
\arguments{
\item{base.cov}{variance/covariance matrix}

\item{...}{ignored}
}
\value{
A list with several fit indexes
\describe{
\item{\code{r.squared.xy}}{\eqn{R^2_{XY}}}
\item{\code{p.squared.yx}}{\eqn{P^2_{YX}}}
}
See \code{\link{mlmWithCov}}
}
\description{
Provides coefficient of determination for multivariate models.
}
\references{
Azen, R., & Budescu, D. V. (2006). Comparing Predictors in Multivariate Regression Models: An Extension of Dominance Analysis. Journal of Educational and Behavioral Statistics, 31(2), 157-180. doi:10.3102/10769986031002157
}
\seealso{
Other fit indexes: \code{\link{da.glm.fit}},
  \code{\link{da.lm.fit}}, \code{\link{da.lmWithCov.fit}},
  \code{\link{da.lmerMod.fit}}
}
\concept{fit indexes}
