% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daAverageContributionByLevel.r
\name{daAverageContributionByLevel}
\alias{daAverageContributionByLevel}
\title{Returns the average contribution by level
presented usually as row between contributions by predictors}
\usage{
daAverageContributionByLevel(x)
}
\arguments{
\item{x}{daRawResults object}
}
\value{
list, with key named as a fit index
              and values are matrix, with the average
              contribution of each variable on every level
}
\description{
Returns the average contribution by level
presented usually as row between contributions by predictors
}
\keyword{internal}

