\name{daAverageContributionByLevel}
\alias{daAverageContributionByLevel}
\title{Returns the average contribution by level
presented usually as row between contributions by predictors}
\usage{
  daAverageContributionByLevel(x)
}
\arguments{
  \item{x}{daRawResults object}
}
\description{
  Returns the average contribution by level presented
  usually as row between contributions by predictors
}

