% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daConditionalDominance.r
\name{daConditionalDominance}
\alias{daConditionalDominance}
\title{Matrix of Conditional dominance of one variable over another.}
\usage{
daConditionalDominance(daRR)
}
\description{
Uses daRawResults as input
}
\seealso{
Other dominance matrices: \code{\link{daCompleteDominance}},
  \code{\link{daGeneralDominance}}
}
\keyword{internal}

