% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daRawResults.r
\name{daRawResults}
\alias{daRawResults}
\title{Retrieve raw results for dominance analysis
Provides name functions, base fit values and 
matrix for models vs predictors importance}
\usage{
daRawResults(x, constants = c(), fit.functions = "default", data = NULL,
  null.model = NULL, ...)
}
\description{
Retrieve raw results for dominance analysis
Provides name functions, base fit values and 
matrix for models vs predictors importance
}

