% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daSubmodels.r
\name{daSubmodels}
\alias{daSubmodels}
\title{Returns all the submodels derived from full models.
You could set some variables as constantes, limiting the number of models.
Includes, by default, the null model}
\usage{
daSubmodels(x, constants = NULL)
}
\arguments{
\item{x}{regression class (lm or lmer)}

\item{constants}{vector of constants}
}
\value{
list with elements level, pred.matrix, predictors, response, constants
}
\description{
Returns all the submodels derived from full models.
You could set some variables as constantes, limiting the number of models.
Includes, by default, the null model
}
\keyword{internal}
