% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominanceMatrix.r
\name{dominanceMatrix}
\alias{dominanceMatrix}
\title{Retrieve or calculates a dominance matrix for a given object}
\usage{
dominanceMatrix(x, undefined.value = 0.5, type = NULL,
  fit.function = NULL, ...)
}
\arguments{
\item{x}{matrix (calculate) or dominanceAnalysis (retrieve)}

\item{undefined.value}{value when no dominance can be established}

\item{type}{type of dominance matrix to retrieve. Could be complete, conditional or general}

\item{fit.function}{specific fit function to retrieve. If not specified, retrieve the first
method according to fit.functions}

\item{...}{extra arguments. Check  \code{\link{dominanceMatrix.matrix}} and \code{\link{dominanceMatrix.dominanceAnalysis}} for specific arguments}
}
\value{
A matrix representing dominance. 1 represented domination of the row variable
        over the column variable, 0 dominance of the column over the row variable.
        Undefined dominance is represented by \code{undefined.value} parameter
}
\description{
This methods calculates or retrieve dominance matrix
}
\details{
To calculate a dominance matrix from a matrix or dataframe, use

\code{dominanceMatrix(x,undefined.value)}.
To retrieve the dominance matrices from a dominanceAnalysis object, use

\code{dominanceMatrix(x,type,fit.function)}
}
