% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daSubmodels.r
\name{formulas.daSubmodels}
\alias{formulas.daSubmodels}
\title{Return a list with formulas for a given daSubmodels object}
\usage{
formulas.daSubmodels(x, env = parent.frame())
}
\arguments{
\item{x}{daSubmodels}

\item{env}{environment}
}
\value{
list
}
\description{
Return a list with formulas for a given daSubmodels object
}
\keyword{internal}
