% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmWithCov.r
\name{lmWithCov}
\alias{lmWithCov}
\title{Uses covariance/correlation matrix for OLS}
\usage{
lmWithCov(f, x)
}
\arguments{
\item{f}{formula}

\item{x}{correlation/covariance matrix}
}
\value{
\item{coef}{coefficients of the regression}
\item{r.squared}{\eqn{R^2} of the regression}
\item{formula}{formula provided as parameter}
\item{cov}{covariance/correlation matrix provided as parameter}
}
\description{
We use for testing, comparing the results of papers with
this package, but useful anyway
}
\examples{
cov.m<-matrix(c(1,0.2,0.3, 0.2,1,0.5,0.3,0.5,1),3,3,
dimnames=list(c("x1","x2","y"),c("x1","x2","y")))
lmWithCov(y~x1+x2,cov.m)
}

