% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bootAverageDominanceAnalysis.r
\name{summary.bootAverageDominanceAnalysis}
\alias{summary.bootAverageDominanceAnalysis}
\title{Summary for bootAverageDominanceAnalysis.}
\usage{
\method{summary}{bootAverageDominanceAnalysis}(object,
  fit.functions = NULL, ...)
}
\arguments{
\item{object}{a \code{\link{bootAverageDominanceAnalysis}} object}

\item{fit.functions}{name of the fit indeces to retrieve. If NULL, all fit indeces will be retrieved}

\item{...}{ignored}
}
\description{
Summary for bootAverageDominanceAnalysis.
}
\keyword{internal}
