% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bootAverageDominanceAnalysis.r
\name{summary.bootAverageDominanceAnalysis}
\alias{summary.bootAverageDominanceAnalysis}
\title{Summary for bootAverageDominanceAnalysis.}
\usage{
\method{summary}{bootAverageDominanceAnalysis}(object, ...)
}
\description{
Summary for bootAverageDominanceAnalysis.
}
