% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bootDominanceAnalysis.r
\name{summary.bootDominanceAnalysis}
\alias{summary.bootDominanceAnalysis}
\title{Summary for bootDominanceAnalysis.}
\usage{
\method{summary}{bootDominanceAnalysis}(x, ...)
}
\description{
Summary for bootDominanceAnalysis.
}

