% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dominanceAnalysis.r
\name{summary.dominanceAnalysis}
\alias{summary.dominanceAnalysis}
\title{Summary for a \code{\link{dominanceAnalysis}} object}
\usage{
\method{summary}{dominanceAnalysis}(object, ...)
}
\arguments{
\item{object}{a \code{\link{dominanceAnalysis}} object}

\item{...}{unused}
}
\value{
A list with values:
\itemize{
\item \code{average.contribution}: vector of average contributions of each variable
\item \code{summary.matrix}: matrix with all calculations for dominance analysis
}
}
\description{
Summary for a \code{\link{dominanceAnalysis}} object
}
\keyword{internal}
