% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmmR2.r
\name{summary.lmmR2}
\alias{summary.lmmR2}
\title{Summary for lmmR2 models}
\usage{
\method{summary}{lmmR2}(object, ...)
}
\arguments{
\item{object}{lmmR2 object}

\item{...}{unused}
}
\value{
An object of class \code{"summary.lmmR2"} containing:
\item{m1}{A data frame with variance information and pseudo R-squared values}
\item{m2}{A data frame with index information, meaning and values}
}
\description{
Summary for lmmR2 models
}
\keyword{internal}
