% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daFitFunctions.r
\name{using-fit-indexes}
\alias{using-fit-indexes}
\title{Provides fit indexes for different regression models.}
\arguments{
\item{data}{complete data}

\item{null.model}{needed for LMM models}

\item{base.cov}{needed when only covarince/correlation matrix is provided}

\item{family.glm}{family for glm models}
}
\value{
function. See \link{using-fit-indexes} description for interface
}
\description{
\code{dominanceAnalysis} try to infer, based on the class of the
model provided, the appropiate fit indexes, using the scheme
da.CLASS.fit for name. This method have two interfaces, one for
retreive the names of fit indexes, and other to retrieve the
indexes based on data.
}
\details{
Interfaces are
\itemize{
\item \code{da.CLASS.fit("names")} returns a vector with names for fit Indexes
\item \code{da.CLASS.fit(data, null.model, base.cov=NULL,family.glm=NULL)} returns a function with one param,
      the formula to calculate the submodel.
}
\code{\link{daRawResults}}
}

